+++ homedocs/server/scripts/buildCssDev.ts
@@ -1,0 +1,3 @@
+import { buildCss } from "./tailwind";
+
+await buildCss()