+++ homedocs/CHANGELOG.md
@@ -1,6 +1,12 @@
-<!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
+## 3.0.1
+
+- migration to nodejs 22 with hono server
 
 ## 2.0.6
+
+- fix run command for bun server
+
+## 2.0.5
 
 - switch base image to debian